+++ src/libs/blueprint/conduit_blueprint_mesh_examples.cpp
@@ -66,7 +66,6 @@ namespace examples
 //---------------------------------------------------------------------------//
 const float64 PI_VALUE = 3.14159265359;
 
-
 //---------------------------------------------------------------------------//
 struct point
 {
@@ -1624,9 +1623,13 @@ braid_quads(index_t npts_x,
 //---------------------------------------------------------------------------//
 void
 braid_quads_and_tris(index_t npts_x,
-            index_t npts_y,
-            Node &res)
+                     index_t npts_y,
+                     Node &res)
 {
+    CONDUIT_INFO("This braid example is deprecated in favor of mixed_2d.")
+
+    // TODO remove in conduit 0.9.4
+
     res.reset();
 
     int32 nele_x = (int32)(npts_x - 1);
@@ -1675,7 +1678,6 @@ braid_quads_and_tris(index_t npts_x,
              ++ielem;
 
         } // END for all i
-
     } // END for all j
 
 
@@ -1744,12 +1746,16 @@ braid_quads_and_tris(index_t npts_x,
     //                                         fields["radial"]);
 }
 
+
 //---------------------------------------------------------------------------//
 void
 braid_quads_and_tris_offsets(index_t npts_x,
                              index_t npts_y,
                              Node &res)
 {
+    CONDUIT_INFO("This braid example is deprecated in favor of mixed_2d.")
+
+    // TODO remove in conduit 0.9.4
 
     res.reset();
 
@@ -2650,6 +2656,9 @@ braid_hexs_and_tets(index_t npts_x,
                     index_t npts_z,
                     Node &res)
 {
+    CONDUIT_INFO("This braid example is deprecated in favor of mixed.")
+
+    // TODO remove in conduit 0.9.4
 
     // WARNING -- The code below is UNTESTED.
     //            The SILO writer is missing an implementation for